common: refactors the C99 API for NS(TrackJobBase) out into common/track_job.h
- The whole C99 API originating from sixtrack::TrackJobBase and which is
  shared along all specializations is factored out into a top-level
  header file common/track_job.h .  Thus, it's not necessary to expose
  C API located in common/internal.

- The C++ API is still located in common/internal/track_job_base.h and is
  threfore exposed; however, since the users are expected to use this API only
  via specialized classes, this should not be too much of a problem.

- TODO: Provide top-level abstractions for all architectures to
  allow tracking, collecting, creation and destruction via the
  generic NS(TrackJob_*) interface (at a small run-time penalty
  due to the virtual fucntion calls and the lack of context to properly
  de-virtualize the instances)
