+++ sixtracklib/common/CMakeLists.txt
@@ -33,6 +33,7 @@ set(  SIXTRACKLIB_C99_HEADERS
       definitions.h
       particles.h
       track.h
+      track_job.h
       track_job_cpu.h
 )
