+++ sixtracklib/common/internal/CMakeLists.txt
@@ -46,6 +46,7 @@ set( SIXTRL_COMMON_INTERNAL_CXX_HEADERS
 
 set( SIXTRL_COMMON_INTERNAL_CXX_SOURCES
      track_job_base.cpp
+     track_job.cpp
      track_job_cpu.cpp
 )
