+++ sixtracklib/common/internal/track_job.cpp
@@ -1,241 +1,403 @@
 #include "sixtracklib/common/track_job.h"
 
-#if !defined( SIXTRL_NO_INCLUDES )
-    #include "sixtracklib/common/buffer.h"
-    #include "sixtracklib/common/internal/track_job_base.h"
-    #include "sixtracklib/common/internal/track_job_cpu.h"
-    #include "sixtracklib/common/output/elem_by_elem_config.h"
-    #include "sixtracklib/common/generated/modules.h"
-
-    #if defined( SIXTRACKLIB_ENABLE_MODULE_OPENCL ) && \
-        SIXTRACKLIB_ENABLE_MODULE_OPENCL == 1
-
-    #include "sixtracklib/opencl/track_job.h"
-
-    #endif /* OpenCL 1.x */
-
-#endif /* !defined( SIXTRL_NO_INCLUDES ) */
-
-#if !defined( _GPUCODE )
 
 #if !defined( SIXTRL_NO_SYSTEM_INCLUDES )
-    #include <cstddef>
-    #include <cstdint>
-    #include <cstdlib>
-    #include <cstring>
-    #include <string>
-    #include <algorithm>
+    #if defined( __cplusplus )
+        #include <cstddef>
+        #include <cstdint>
+        #include <cstdlib>
+        #include <memory>
+        #include <string>
+        #include <vector>
+    #else /* !defined( __cplusplus ) */
+        #include <stdbool.h>
+        #include <stddef.h>
+        #include <stdint.h>
+        #include <stdlib.h>
+        #include <limits.h>
+    #endif /* !defined( __cplusplus ) */
 #endif /* !defined( SIXTRL_NO_SYSTEM_INCLUDES ) */
 
 #if !defined( SIXTRL_NO_INCLUDES )
-    #include "sixtracklib/common/buffer.hpp"
+    #include "sixtracklib/common/buffer.h"
+    #include "sixtracklib/common/definitions.h"
+    #include "sixtracklib/common/particles.h"
+    #include "sixtracklib/common/beam_elements.h"
+    #include "sixtracklib/common/be_monitor/be_monitor.h"
+    #include "sixtracklib/common/context/context_abs_base.h"
+    #include "sixtracklib/common/output/elem_by_elem_config.h"
+    #include "sixtracklib/common/output/output_buffer.h"
+    #include "sixtracklib/common/internal/track_job_base.h"
 #endif /* !defined( SIXTRL_NO_INCLUDES ) */
 
-namespace SIXTRL_CXX_NAMESPACE
+#if !defined( _GPUCODE ) && ( defined( __cplusplus ) )
+
+SIXTRL_HOST_FN void NS(TrackJob_clear)( NS(TrackJobBase)* SIXTRL_RESTRICT job )
 {
-    SIXTRL_HOST_FN TrackJobBase* createTrackJob(
-        const char *const SIXTRL_RESTRICT type_str,
-        const char *const SIXTRL_RESTRICT device_id_str,
-        const char *const SIXTRL_RESTRICT config_str )
-    {
-        using size_t = Buffer::size_type;
-
-        std::string const _type_str =
-            ( ( type_str != nullptr ) &&
-              ( std::strlen( type_str ) > size_t{ 0 } ) )
-            ? std::string{ type_str } : std::string{};
-
-        std::string const _device_id =
-            ( ( device_id_str != nullptr ) &&
-              ( std::strlen( device_id_str ) > size_t{ 0 } ) )
-                ? std::string{ device_id_str } : std::string{};
-
-        std::string const _config_str =
-            ( ( config_str != nullptr ) &&
-              ( std::strlen( config_str ) > size_t{ 0 } ) )
-                ? std::string{ config_str } : std::string{};
-
-        return createTrackJob( _type_str, _device_id, _config_str );
-    }
-
-    SIXTRL_HOST_FN TrackJobBase* createTrackJob(
-        const char *const SIXTRL_RESTRICT type_str,
-        ::NS(Buffer)* particles_buffer,
-        ::NS(Buffer)* beam_elements_buffer,
-        const char *const SIXTRL_RESTRICT device_id_str,
-        const char *const SIXTRL_RESTRICT config_str,
-        ::NS(buffer_size_t) const num_elem_by_elem_turns )
-    {
-        TrackJobBase* ptr_job = createTrackJob(
-            type_str, device_id_str, config_str );
-
-        if( ptr_job != nullptr )
-        {
-            ptr_job->reset( particles_buffer, beam_elements_buffer, nullptr,
-                            num_elem_by_elem_turns );
-        }
-
-        return ptr_job;
-    }
-
-    SIXTRL_HOST_FN TrackJobBase* createTrackJob(
-        const char *const SIXTRL_RESTRICT type_str,
-        ::NS(Buffer)* particles_buffer,
-        ::NS(Buffer)* beam_elements_buffer,
-        ::NS(Buffer)* output_buffer,
-        const char *const SIXTRL_RESTRICT device_id_str,
-        const char *const SIXTRL_RESTRICT config_str,
-        ::NS(buffer_size_t) const num_elem_by_elem_turns )
-    {
-        TrackJobBase* ptr_job = createTrackJob(
-            type_str, device_id_str, config_str );
-
-        if( ptr_job != nullptr )
-        {
-            ptr_job->reset( particles_buffer, beam_elements_buffer,
-                            output_buffer, num_elem_by_elem_turns );
-        }
-
-        return ptr_job;
-    }
-
-
-    SIXTRL_HOST_FN TrackJobBase* createTrackJob(
-        std::string const& SIXTRL_RESTRICT_REF type_str,
-        std::string const& SIXTRL_RESTRICT_REF device_id_str,
-        std::string const& SIXTRL_RESTRICT_REF config_str )
-    {
-        TrackJobBase* ptr_job = nullptr;
-
-        if( 0 == type_str.compare( SIXTRL_CXX_NAMESPACE::TRACK_JOB_CPU_STR ) )
-        {
-            ptr_job = new SIXTRL_CXX_NAMESPACE::TrackJobCpu( config_str );
-        }
-        #if defined( SIXTRACKLIB_ENABLE_MODULE_OPENCL ) && \
-            SIXTRACKLIB_ENABLE_MODULE_OPENCL == 1
-        else if( 0 == type_str.compare( "opencl" ) )
-        {
-            ptr_job = new SIXTRL_CXX_NAMESPACE::TrackJobOcl(
-                device_id, config_str );
-        }
-        #endif /* OpenCL 1.x Module */
-
-        return ptr_job;
-    }
-
-    SIXTRL_HOST_FN TrackJobBase* createTrackJob(
-        std::string const& SIXTRL_RESTRICT type_str,
-        Buffer& particles_buffer,
-        Buffer& beam_elements_buffer,
-        std::string const& SIXTRL_RESTRICT device_id_str,
-        std::string const& SIXTRL_RESTRICT config_str,
-        Buffer::size_type const num_elem_by_elem_turns )
-    {
-        TrackJobBase* ptr_job = createTrackJob( type_str,
-            device_id_str, config_str );
-
-        if( ptr_job != nullptr )
-        {
-            ptr_job->reset( particles_buffer, beam_elements_buffer,
-                            num_elem_by_elem_turns );
-        }
-
-        return ptr_job;
-    }
-
-    SIXTRL_HOST_FN TrackJobBase* createTrackJob(
-        std::string const& SIXTRL_RESTRICT type_str,
-        Buffer& particles_buffer,
-        Buffer& beam_elements_buffer,
-        Buffer& output_buffer,
-        std::string const& SIXTRL_RESTRICT device_id_str,
-        std::string const& SIXTRL_RESTRICT config_str,
-        Buffer::size_type const num_elem_by_elem_turns )
-    {
-        TrackJobBase* ptr_job = createTrackJob( type_str,
-            device_id_str, config_str );
-
-        if( ptr_job != nullptr )
-        {
-            ptr_job->reset( particles_buffer, beam_elements_buffer,
-                            output_buffer, num_elem_by_elem_turns );
-        }
-
-        return ptr_job;
-    }
+    if( job != nullptr ) job->clear();
 }
 
-SIXTRL_HOST_FN ::NS(TrackJobBase)* NS(TrackJob_new)(
-    const char *const SIXTRL_RESTRICT type_str,
-    const char *const SIXTRL_RESTRICT device_id_str,
-    ::NS(Buffer)* SIXTRL_RESTRICT particles_buffer,
-    ::NS(Buffer)* SIXTRL_RESTRICT beam_elements_buffer,
-    ::NS(Buffer)* SIXTRL_RESTRICT output_buffer,
-    const char *const SIXTRL_RESTRICT config_str )
+SIXTRL_HOST_FN void NS(TrackJob_collect)( NS(TrackJobBase)* SIXTRL_RESTRICT j )
 {
-    return SIXTRL_CXX_NAMESPACE::createTrackJob(
-        type_str, device_id_str, particles_buffer,
-            beam_elements_buffer, output_buffer, config_str );
+    if( j != nullptr ) j->collect();
 }
 
-SIXTRL_HOST_FN void NS(TrackJob_track_until_turn)(
-    ::NS(TrackJobBase)* SIXTRL_RESTRICT job,
-    ::NS(Buffer)* SIXTRL_RESTRICT particles_buffer,
-    ::NS(Buffer)* SIXTRL_RESTRICT beam_elements_buffer,
-    ::NS(buffer_size_t) const until_turn )
-{
-    if( job != nullptr )
-    {
-        job->track( particles_buffer, beam_elements_buffer, until_turn );
-    }
-
-    return;
-}
-
-SIXTRL_HOST_FN void NS(TrackJob_track_elem_by_elem)(
-    ::NS(TrackJobBase)* SIXTRL_RESTRICT job,
-    ::NS(Buffer)* SIXTRL_RESTRICT particles_buffer,
-    ::NS(Buffer)* SIXTRL_RESTRICT beam_elements_buffer,
-    ::NS(buffer_size_t) const num_elem_by_elem_turns )
-{
-    if( job != nullptr )
-    {
-        job->trackElemByElem(
-            particles_buffer, beam_elements_buffer,
-            num_elem_by_elem_turns );
-    }
-
-    return;
-}
-
-SIXTRL_HOST_FN void NS(TrackJob_delete)(
-    ::NS(TrackJobBase)* SIXTRL_RESTRICT job )
-{
-    if( job != nullptr ) delete job;
-    return;
-}
-
-SIXTRL_HOST_FN ::NS(Buffer)* NS(TrackJob_get_output_buffer)(
-    const ::NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+SIXTRL_HOST_FN bool NS(TrackJob_reset)( NS(TrackJobBase)* SIXTRL_RESTRICT job,
+    NS(Buffer)* SIXTRL_RESTRICT particles_buffer,
+    NS(Buffer)* SIXTRL_RESTRICT beam_elem_buffer,
+    NS(Buffer)* SIXTRL_RESTRICT ptr_output_buffer )
 {
     return ( job != nullptr )
-        ? job->ptrOutputBuffer() : nullptr;
+        ? job->reset( particles_buffer, beam_elem_buffer, ptr_output_buffer )
+        : false;
 }
 
-SIXTRL_HOST_FN ::NS(ElemByElemConfig)* ::NS(TrackJob_get_elem_by_elem_config)(
-    const ::NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+SIXTRL_HOST_FN bool NS(TrackJob_reset_with_output)(
+    NS(TrackJobBase)* SIXTRL_RESTRICT job,
+    NS(Buffer)* SIXTRL_RESTRICT particles_buffer,
+    NS(Buffer)* SIXTRL_RESTRICT beam_elem_buffer,
+    NS(Buffer)* SIXTRL_RESTRICT output_buffer,
+    NS(buffer_size_t) const dump_elem_by_elem_turns )
+{
+    return ( job != nullptr )
+        ? job->reset( particles_buffer, beam_elem_buffer, output_buffer,
+                      dump_elem_by_elem_turns )
+        : false;
+}
+
+SIXTRL_HOST_FN bool NS(TrackJob_reset_detailed)(
+    NS(TrackJobBase)* SIXTRL_RESTRICT job,
+    NS(Buffer)* SIXTRL_RESTRICT particles_buffer,
+    NS(buffer_size_t) const num_particle_sets,
+    NS(buffer_size_t) const* SIXTRL_RESTRICT particle_set_indices_begin,
+    NS(Buffer)* SIXTRL_RESTRICT beam_elem_buffer,
+    NS(Buffer)* SIXTRL_RESTRICT output_buffer,
+    NS(buffer_size_t) const dump_elem_by_elem_turns )
+{
+    return ( job != nullptr ) ? job->reset( particles_buffer, num_particle_sets,
+            particle_set_indices_begin, beam_elem_buffer, output_buffer,
+                dump_elem_by_elem_turns )
+        : false;
+}
+
+SIXTRL_HOST_FN bool NS(TrackJob_assign_output_buffer)(
+    NS(TrackJobBase)* SIXTRL_RESTRICT job,
+    NS(Buffer)* SIXTRL_RESTRICT ptr_output_buffer )
+{
+    return ( job != nullptr )
+        ? job->assignOutputBuffer( ptr_output_buffer ) : false;
+}
+
+/* ------------------------------------------------------------------------- */
+
+SIXTRL_HOST_FN NS(track_job_type_t) NS(TrackJob_get_type_id)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->type() : ::NS(track_job_type_t){ -1 };
+}
+
+SIXTRL_HOST_FN char const* NS(TrackJob_get_type_str)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->typeStr().c_str() : nullptr;
+}
+
+SIXTRL_HOST_FN bool NS(TrackJob_has_device_id_str)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->hasDeviceIdStr() : false;
+}
+
+SIXTRL_HOST_FN char const* NS(TrackJob_get_device_id_str)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->deviceIdStr().c_str() : nullptr;
+}
+
+SIXTRL_HOST_FN bool NS(TrackJob_has_config_str)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->hasConfigStr() : false;
+}
+
+SIXTRL_HOST_FN char const* NS(TrackJob_get_config_str)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->configStr().c_str() : nullptr;
+}
+
+/* ------------------------------------------------------------------------- */
+
+SIXTRL_HOST_FN NS(buffer_size_t) NS(TrackJob_get_num_particle_sets)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr )
+        ? job->numParticleSets() : ::NS(buffer_size_t){ 0 };
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t) const*
+NS(TrackJob_get_particle_set_indices_begin)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->particleSetIndicesBegin() : nullptr;
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t) const*
+NS(TrackJob_get_particle_set_indices_end)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->particleSetIndicesEnd() : nullptr;
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t) NS(TrackJob_get_particle_set_index)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job,
+    NS(buffer_size_t) const n )
+{
+    SIXTRL_ASSERT( job != nullptr );
+    return job->particleSetIndex( n );
+}
+
+/* ------------------------------------------------------------------------- */
+
+SIXTRL_HOST_FN NS(particle_index_t) NS(TrackJob_get_min_particle_id)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    using index_t = ::NS(particle_index_t);
+    return ( job != nullptr ) ? job->minParticleId() : index_t{ -1 };
+}
+
+SIXTRL_HOST_FN NS(particle_index_t) NS(TrackJob_get_max_particle_id)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    using index_t = ::NS(particle_index_t);
+    return ( job != nullptr ) ? job->maxParticleId() : index_t{ -1 };
+}
+
+SIXTRL_HOST_FN NS(particle_index_t) NS(TrackJob_get_min_element_id)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    using index_t = ::NS(particle_index_t);
+    return ( job != nullptr ) ? job->minElementId() : index_t{ -1 };
+}
+
+SIXTRL_HOST_FN NS(particle_index_t) NS(TrackJob_get_max_element_id)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    using index_t = ::NS(particle_index_t);
+    return ( job != nullptr ) ? job->maxElementId() : index_t{ -1 };
+}
+
+SIXTRL_HOST_FN NS(particle_index_t) NS(TrackJob_get_min_initial_turn_id)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    using index_t = ::NS(particle_index_t);
+    return ( job != nullptr ) ? job->minInitialTurnId() : index_t{ -1 };
+}
+
+SIXTRL_HOST_FN NS(particle_index_t) NS(TrackJob_get_max_initial_turn_id)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    using index_t = ::NS(particle_index_t);
+    return ( job != nullptr ) ? job->maxInitialTurnId() : index_t{ -1 };
+}
+
+/* ------------------------------------------------------------------------- */
+
+SIXTRL_HOST_FN NS(Buffer)* NS(TrackJob_get_particles_buffer)(
+    NS(TrackJobBase)* SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->ptrCParticlesBuffer() : nullptr;
+}
+
+SIXTRL_HOST_FN NS(Buffer) const* NS(TrackJob_get_const_particles_buffer)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->ptrCParticlesBuffer() : nullptr;
+}
+
+SIXTRL_HOST_FN NS(Buffer)*
+NS(TrackJob_get_beam_elements_buffer)( NS(TrackJobBase)* SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->ptrCBeamElementsBuffer() : nullptr;
+}
+
+SIXTRL_HOST_FN NS(Buffer) const*
+NS(TrackJob_get_const_beam_elements_buffer)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->ptrCBeamElementsBuffer() : nullptr;
+}
+
+/* ------------------------------------------------------------------------- */
+
+SIXTRL_HOST_FN bool NS(TrackJob_has_output_buffer)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->hasOutputBuffer() : false;
+}
+
+SIXTRL_HOST_FN bool NS(TrackJob_owns_output_buffer)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->ownsOutputBuffer() : false;
+}
+
+SIXTRL_HOST_FN bool NS(TrackJob_has_elem_by_elem_output)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->hasElemByElemOutput() : false;
+}
+
+SIXTRL_HOST_FN bool NS(TrackJob_has_beam_monitor_output)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->hasBeamMonitorOutput() : false;
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t)
+NS(TrackJob_get_beam_monitor_output_buffer_offset)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    SIXTRL_ASSERT( job != nullptr );
+    return job->beamMonitorsOutputBufferOffset();
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t)
+NS(TrackJob_get_elem_by_elem_output_buffer_offset)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    SIXTRL_ASSERT( job != nullptr );
+    return job->elemByElemOutputBufferOffset();
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t)
+NS(TrackJob_get_num_elem_by_elem_turns)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr )
+        ? job->numElemByElemTurns() : ::NS(buffer_size_t){ 0 };
+}
+
+SIXTRL_HOST_FN NS(Buffer)* NS(TrackJob_get_output_buffer)(
+    NS(TrackJobBase)* SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->ptrCOutputBuffer() : nullptr;
+}
+
+SIXTRL_HOST_FN NS(Buffer) const*
+NS(TrackJob_get_const_output_buffer)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->ptrCOutputBuffer() : nullptr;
+}
+
+/* ------------------------------------------------------------------------- */
+
+SIXTRL_HOST_FN bool NS(TrackJob_has_beam_monitors)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->hasBeamMonitors() : false;
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t) NS(TrackJob_get_num_beam_monitors)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr )
+        ? job->numBeamMonitors() : ::NS(buffer_size_t){ 0 };
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t) const*
+NS(TrackJob_get_beam_monitor_indices_begin)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->beamMonitorIndicesBegin() : nullptr;
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t) const*
+NS(TrackJob_get_beam_monitor_indices_end)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->beamMonitorIndicesEnd() : nullptr;
+}
+
+SIXTRL_HOST_FN NS(buffer_size_t) NS(TrackJob_get_beam_monitor_index)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job,
+    NS(buffer_size_t) const n )
+{
+    return ( job != nullptr )
+        ? job->beamMonitorIndex( n ) : ::NS(buffer_size_t){ 0 };
+}
+
+/* ------------------------------------------------------------------------- */
+
+SIXTRL_HOST_FN bool NS(TrackJob_has_elem_by_elem_config)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->hasElemByElemConfig() : false;
+}
+
+SIXTRL_HOST_FN NS(ElemByElemConfig) const*
+NS(TrackJob_get_elem_by_elem_config)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->ptrElemByElemConfig() : nullptr;
+}
+
+SIXTRL_HOST_FN bool NS(TrackJob_is_elem_by_elem_config_rolling)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->elemByElemRolling() : false;
+}
+
+SIXTRL_HOST_FN bool
+NS(TrackJob_get_default_elem_by_elem_config_rolling_flag)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr ) ? job->defaultElemByElemRolling() : false;
+}
+
+SIXTRL_HOST_FN void
+NS(TrackJob_set_default_elem_by_elem_config_rolling_flag)(
+    NS(TrackJobBase)* SIXTRL_RESTRICT job, bool const is_rolling_flag )
 {
     if( job != nullptr )
     {
-        return job->elemByElemConfig();
+        job->setDefaultElemByElemRolling( is_rolling_flag );
     }
-    else
-    {
-        ::NS(ElemByElemConfig) dummy;
-        ::NS(ElemByElemConfig_preset)( &dummy );
 
-        return dummy;
-    }
+    return;
 }
+
+SIXTRL_HOST_FN NS(elem_by_elem_order_t)
+NS(TrackJob_get_elem_by_elem_config_order)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr )
+        ? job->elemByElemOrder() : ::NS(ELEM_BY_ELEM_ORDER_INVALID);
+}
+
+SIXTRL_HOST_FN NS(elem_by_elem_order_t)
+NS(TrackJob_get_default_elem_by_elem_config_order)(
+    const NS(TrackJobBase) *const SIXTRL_RESTRICT job )
+{
+    return ( job != nullptr )
+        ? job->defaultElemByElemOrder()
+        : ::NS(ELEM_BY_ELEM_ORDER_INVALID);
+}
+
+SIXTRL_HOST_FN void
+NS(TrackJob_set_default_elem_by_elem_config_order)(
+    NS(TrackJobBase)* SIXTRL_RESTRICT job,
+    NS(elem_by_elem_order_t) const order )
+{
+    if( job != nullptr )
+    {
+        job->setDefaultElemByElemOrder( order );
+    }
+
+    return;
+}
+
+#endif /* !defined( _GPUCODE ) && ( defined( __cplusplus ) ) */
 
 /* end: sixtracklib/common/internal/track_job.cpp */
