+++ sixtracklib/sixtracklib.h
@@ -46,6 +46,7 @@
 #include "sixtracklib/common/constants.h"
 #include "sixtracklib/common/buffer.h"
 #include "sixtracklib/common/particles.h"
+#include "sixtracklib/common/track_job.h"
 #include "sixtracklib/common/track_job_cpu.h"
 #include "sixtracklib/common/track.h"
